Seperate scalatest for tokenization. reverted all tokenization related changes and tests from ColumnTableTest.
