+++ core/src/test/scala/org/apache/spark/sql/store/ColumnTableTest.scala
@@ -20,20 +20,22 @@ import java.sql.DriverManager
 import java.util
 
 import scala.util.{Failure, Success, Try}
+
 import com.gemstone.gemfire.cache.{EvictionAction, EvictionAlgorithm}
 import com.gemstone.gemfire.internal.cache.{GemFireCacheImpl, PartitionedRegion}
 import com.pivotal.gemfirexd.internal.engine.Misc
 import com.pivotal.gemfirexd.internal.impl.jdbc.EmbedConnection
 import com.pivotal.gemfirexd.internal.impl.sql.compile.ParserImpl
-import io.snappydata.{SnappyFunSuite, SnappyTableStatsProviderService}
+import io.snappydata.{SnappyTableStatsProviderService, SnappyFunSuite}
 import io.snappydata.core.{Data, TestData, TestData2}
 import org.apache.hadoop.hive.ql.parse.ParseDriver
 import org.scalatest.{BeforeAndAfter, BeforeAndAfterAll}
+
 import org.apache.spark.Logging
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
 import org.apache.spark.sql.execution.columnar.JDBCAppendableRelation
 import org.apache.spark.sql.types.{IntegerType, StructField, StructType}
-import org.apache.spark.sql._
+import org.apache.spark.sql.{AnalysisException, DataFrame, SaveMode, SparkSession, TableNotFoundException}
 
 /**
   * Tests for column tables in GFXD.
@@ -72,7 +74,7 @@ class ColumnTableTest
       tableName: String, df: DataFrame, startCount: Int, size: Int): Int = {
 
     var count = startCount
-    var result = snc.sql(s"SELECT * FROM $schemaName.$tableName where a = 1")
+    var result = snc.sql(s"SELECT * FROM $schemaName.$tableName")
     assert(result.collect().length === count)
 
     snc.sql(String.format(pattern, schemaName))
@@ -139,106 +141,6 @@ class ColumnTableTest
 
   test("Test the creation/dropping of column table using Schema") {
     val data = Seq(Seq(1, 2, 3), Seq(7, 8, 9), Seq(9, 2, 3),
-      Seq(4, 2, 4), Seq(5, 6, 7))
-    val rdd = sc.parallelize(data, data.length)
-        .map(s => Data(s.head, s(1), s(2)))
-    val dataDF = snc.createDataFrame(rdd)
-
-    val schema = "test"
-    val table = "MY_TABLE"
-
-    snc.sql(s"Drop Table if exists $schema.$table")
-    snc.sql(s"Create Table $schema.$table (a INT, b INT, c INT) " +
-        "using column options()")
-
-    // try different variant of set schema
-    val size = dataDF.count().toInt
-    var count = 0
-
-    try {
-      val lp1 = snc.snappySession.onlyParseSQL(s"SELECT * FROM $schema.$table where a = 1")
-      val lp2 = snc.snappySession.onlyParseSQL(s"SELECT * FROM $schema.$table where a = 2")
-      val hc1 = scala.util.hashing.MurmurHash3.productHash(lp1)
-      val hc2 = scala.util.hashing.MurmurHash3.productHash(lp2)
-      println("HC1 = " + hc1 + " and HC2 = " + hc2)
-      assert(lp1.equals(lp2))
-      dataDF.write.insertInto(s"$schema.$table")
-      Thread.sleep(5000)
-      var result = snc.sql(s"SELECT * FROM $schema.$table where a = 1").collect()
-      assert(result.length === 1)
-      assert(result(0).toSeq === data.head)
-
-      result = snc.sql(s"SELECT * FROM $schema.$table where a = 4").collect()
-      assert(result.length === 1)
-      assert(result(0).toSeq === data(3))
-
-      result = snc.sql(s"SELECT * FROM $schema.$table where a = 2").collect()
-      assert(result.length === 0)
-
-      snc.sql(s"drop table $schema.$table")
-    } finally {
-      snc.sql("set spark.sql.caseSensitive = false")
-      snc.sql("set schema = APP")
-    }
-
-    logInfo("Successful")
-  }
-
-  test("Test tokenize") {
-    val numRows = 10000
-    val data = ((0 to numRows), (0 to numRows), (0 to numRows)).zipped.toArray
-
-
-//    val data = Seq(Seq(1, 2, 3), Seq(7, 8, 9), Seq(9, 2, 3),
-//      Seq(4, 2, 4), Seq(5, 6, 7))
-    val rdd = sc.parallelize(data, data.length)
-      .map(s => Data(s._1, s._2, s._3))
-    val dataDF = snc.createDataFrame(rdd)
-
-    val schema = "test"
-    val table = "MY_TABLE"
-
-    snc.sql(s"Drop Table if exists $schema.$table")
-    snc.sql(s"Create Table $schema.$table (a INT, b INT, c INT) " +
-      "using column options()")
-
-    // try different variant of set schema
-    val size = dataDF.count().toInt
-    var count = 0
-
-    try {
-      dataDF.write.insertInto(s"$schema.$table")
-      Thread.sleep(5000)
-
-      //println("query starts")
-      val q = (0 until 1000) map { x =>
-        s"SELECT * FROM $schema.$table where a = $x"
-      }
-      //println("query ends")
-      val start = System.currentTimeMillis()
-      q map { x =>
-        var result = snc.sql(x).collect()
-        assert(result.length === 1)
-        //if ( x % 10 == 0) {
-        //println(s"query = $x and result = " + result(0))
-        //}
-      }
-      val end = System.currentTimeMillis()
-
-      snc.sql(s"SELECT * FROM $schema.$table where a = 1200").collect()
-      println("Time taken = " + (end - start))
-
-      snc.sql(s"drop table $schema.$table")
-    } finally {
-      snc.sql("set spark.sql.caseSensitive = false")
-      snc.sql("set schema = APP")
-    }
-
-    logInfo("Successful")
-  }
-
-  test("Test the creation/dropping 2 of column table using Schema") {
-    val data = Seq(Seq(1, 2, 3), Seq(7, 8, 9), Seq(9, 2, 3),
       Seq(4, 2, 3), Seq(5, 6, 7))
     val rdd = sc.parallelize(data, data.length)
         .map(s => Data(s.head, s(1), s(2)))
@@ -280,6 +182,7 @@ class ColumnTableTest
 
     logInfo("Successful")
   }
+
 
   test("Test the creation/dropping of table using Snappy API") {
     // shouldn't be able to create without schema
